[1341] FastEvaluation version 1.0 (#1417)
* create subdirectories in evaluataion package

* add readme and update version

* lint

* updating evaluation config (#1419)

* fix comments

* lint

---------
