+++ config/evaluate/eval_config.yml
@@ -4,6 +4,7 @@
 #  region: ["belgium", "global"]
 #  image_format : "png" #options: "png", "pdf", "svg", "eps", "jpg" ..
 #  dpi_val : 300
+#  fps: 2
 #  ERA5:
 #    marker_size: 2
 #    scale_marker_size: 1
@@ -27,7 +28,9 @@ evaluation:
   log_scale: false
   add_grid: false
   score_cards: false
+  bar_plots: false
   num_processes: 0  #options: int, "auto", 0 means no parallelism (default)
+  # baseline: "ar40mckx"
 
 run_ids :
   ar40mckx:
@@ -47,6 +50,7 @@ run_ids :
           forecast_step: [1,3, 2]
           ensemble: [0,2,5] #supported: "all", "mean", [0,1,2]
           plot_maps: true
+          plot_target: false
           plot_histograms: true
           plot_animations: true
       CERRA:
@@ -58,6 +62,7 @@ run_ids :
           sample: [2, 3, 0]
           forecast_step: [1,3, 4, 5]
           plot_maps: true
+          plot_target: false
           plot_histograms: true
           plot_animations: true
       
@@ -78,5 +83,6 @@ run_ids :
           forecast_step: [1,3, 2]
           ensemble: "mean" 
           plot_maps: true
+          plot_target: false
           plot_histograms: true
           plot_animations: true
