+++ packages/evaluate/pyproject.toml
@@ -1,8 +1,8 @@
 [project]
 name = "weathergen-evaluate"
-version = "0.1.0"
+version = "1.0"
 description = "The WeatherGenerator Machine Learning Earth System Model"
-readme = "../../README.md"
+readme = "./README.md"
 requires-python = ">=3.12,<3.13"
 dependencies = [
     "cartopy>=0.24.1",
