+++ packages/evaluate/src/weathergen/evaluate/plotting/plot_utils.py
