+++ packages/evaluate/src/weathergen/evaluate/run_evaluation.py
@@ -9,6 +9,7 @@
 # weathergen-evaluate = { path = "../../../../../packages/evaluate" }
 # ///
 
+# Standard library
 import argparse
 import logging
 import multiprocessing as mp
@@ -17,16 +18,19 @@ from collections import defaultdict
 from logging.handlers import QueueHandler, QueueListener
 from pathlib import Path
 
+# Third-party
 import mlflow
 from mlflow.client import MlflowClient
 from omegaconf import DictConfig, OmegaConf
 
+# Local application / package
 from weathergen.common.config import _REPO_ROOT
 from weathergen.common.logger import init_loggers
 from weathergen.common.platform_env import get_platform_env
-from weathergen.evaluate.io_reader import CsvReader, WeatherGenReader
-from weathergen.evaluate.plot_utils import collect_channels
-from weathergen.evaluate.utils import (
+from weathergen.evaluate.io.csv_reader import CsvReader
+from weathergen.evaluate.io.wegen_reader import WeatherGenReader
+from weathergen.evaluate.plotting.plot_utils import collect_channels
+from weathergen.evaluate.utils.utils import (
     calc_scores_per_stream,
     plot_data,
     plot_summary,
